new changes

+++ APP.Eds/APP.Eds/Models/ShoppingProduct/CompartimentResponse.cs
@@ -12,13 +12,13 @@ public class CompartimentResponse
     public int Number { get; set; }
 
     [JsonPropertyName("nominal")]
-    public int Nominal { get; set; }
+    public double Nominal { get; set; }
 
     [JsonPropertyName("operative")]
-    public int Operative { get; set; }
+    public double Operative { get; set; }
 
     [JsonPropertyName("stock")]
-    public int Stock { get; set; }
+    public double Stock { get; set; }
 
     [JsonPropertyName("height")]
     public int Height { get; set; }
